+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Now each podcast has two settings pages (general settings, episode settings)
 - Improved fetch from RSS via CURL 
 - Improved fetch from RSS progress notifications
+- Fixed route for episodes
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 - Fixed permission issues where non-admin users were not able to access some plugin actions